Add email service interface
- Move implementation to EmailServiceImpl
